+++ iphone/Classes/TiGradient.h
@@ -33,7 +33,7 @@ typedef enum {
 	
 	CFMutableArrayRef colorValues;
 	CGFloat * colorOffsets;	//A -1 indicates a lack of entry.
-	int arraySize;
+	NSUInteger arraySize;
 	int offsetsDefined;
 @private
